+++ product/dashboard/src/main/java/io/codekvast/dashboard/file_import/impl/InvocationDataImporterImpl.java
@@ -22,27 +22,27 @@
 package io.codekvast.dashboard.file_import.impl;
 
 import io.codekvast.dashboard.file_import.InvocationDataImporter;
+import io.codekvast.dashboard.metrics.MetricsService;
 import io.codekvast.javaagent.model.v2.CommonPublicationData2;
 import io.codekvast.javaagent.model.v2.InvocationDataPublication2;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.inject.Inject;
 import java.util.TreeSet;
+
+import static io.codekvast.dashboard.metrics.MetricsService.PublicationKind.INVOCATIONS;
 
 /**
  * @author olle.hallin@crisp.se
  */
 @Component
 @Slf4j
+@RequiredArgsConstructor
 public class InvocationDataImporterImpl implements InvocationDataImporter {
     private final ImportDAO importDAO;
-
-    @Inject
-    public InvocationDataImporterImpl(ImportDAO importDAO) {
-        this.importDAO = importDAO;
-    }
+    private final MetricsService metricsService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
@@ -57,6 +57,7 @@ public class InvocationDataImporterImpl implements InvocationDataImporter {
 
         importDAO.importInvocations(customerId, appId, environmentId, jvmId, publication.getRecordingIntervalStartedAtMillis(),
                                     new TreeSet<>(publication.getInvocations()));
+        metricsService.gaugePublicationSize(INVOCATIONS, commonData.getEnvironment(), publication.getInvocations().size());
         return true;
     }
 }