+++ product/dashboard/src/test/java/io/codekvast/dashboard/file_import/impl/PublicationImporterImplTest.java
@@ -3,6 +3,7 @@ package io.codekvast.dashboard.file_import.impl;
 import io.codekvast.dashboard.file_import.CodeBaseImporter;
 import io.codekvast.dashboard.file_import.InvocationDataImporter;
 import io.codekvast.dashboard.file_import.PublicationImporter;
+import io.codekvast.dashboard.metrics.MetricsService;
 import io.codekvast.javaagent.model.v2.CodeBasePublication2;
 import io.codekvast.javaagent.model.v2.InvocationDataPublication2;
 import org.assertj.core.util.Files;
@@ -37,12 +38,15 @@ public class PublicationImporterImplTest {
     @Mock
     private Validator validator;
 
+    @Mock
+    private MetricsService metricsService;
+
     private PublicationImporter publicationImporter;
 
     @Before
     public void beforeTest() {
         MockitoAnnotations.initMocks(this);
-        this.publicationImporter = new PublicationImporterImpl(codeBaseImporter, invocationDataImporter, validator);
+        this.publicationImporter = new PublicationImporterImpl(codeBaseImporter, invocationDataImporter, validator, metricsService);
     }
 
     @Test